# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Solar Installation CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset user password (admin generates temp password)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Temporary password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TempPasswordResult"

  # ─── LEADS ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TempPasswordResult:
      type: object
      properties:
        tempPassword:
          type: string
        message:
          type: string
      required: [tempPassword, message]
    UserRole:
      type: string
      enum: [admin, sales, finance, warehouse, engineer]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, email, role, isActive, createdAt, updatedAt]
    UserInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        phone:
          type: string
        password:
          type: string
        isActive:
          type: boolean
      required: [name, email, role, password]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        phone:
          type: string
        isActive:
          type: boolean
...